Test Card for calibration
Inspired by old television test card, this a simple web page that shows RGB 
colors as well as black and white. They will be used as references shown on a screen
with a decent color quality (i.e my phone) to know the "true" color of what the webcam
sees.