feat: basic optimization implementation
Start by finding the contour of the test card, and get the position of 5
points representing the 5 colors we want to optimize
Then, we use the v4l2-ctl utility to control the webcam
finally we use a means square error between the color we currently have
and the groundtruth and ty to minimize it.
Does not work super well so far.